refactor: mover endpoint de hallazgos de tema a informe para mejor organización REST
