+++ src/informe/informe.service.ts
@@ -5,12 +5,15 @@ import { FilterDto } from '../filters/filters.dto';
 import { FiltersService } from '../filters/filters.service';
 import { Informe } from './schemas/informe.schema';
 import { InformeDTO } from './dto/informe.dto';
+import { Tema } from '../tema/schemas/tema.schema';
 
 @Injectable()
 export class InformeService {
     constructor(
         @InjectModel(Informe.name)
         private readonly InformeModel: Model<Informe>,
+        @InjectModel(Tema.name)
+        private readonly TemaModel: Model<Tema>,
     ) { }
     
     private populateFields(): any[] {
@@ -80,5 +83,45 @@ export class InformeService {
         return await this.InformeModel
             .countDocuments(filtersService.getQuery())
             .exec();
+    }
+
+    async getHallazgosByInforme(informeId: string): Promise<any[]> {
+        const informe = await this.InformeModel.findById(informeId).exec();
+        if (!informe) {
+            throw new Error(`Informe ${informeId} no existe`);
+        }
+
+        const temas = await this.TemaModel
+            .find({ informe_id: informeId, activo: true })
+            .exec();
+
+        const hallazgos = [];
+
+        temas.forEach(tema => {
+            tema.subtema.forEach(subtema => {
+                if (subtema.activo) {
+                    subtema.hallazgo.forEach(hallazgo => {
+                        if (hallazgo.activo) {
+                            hallazgos.push({
+                                _id: hallazgo._id,
+                                titulo: hallazgo.titulo,
+                                criterio: hallazgo.criterio,
+                                descripcion: hallazgo.descripcion,
+                                activo: hallazgo.activo,
+                                tema_id: tema._id,
+                                tema_titulo: tema.titulo,
+                                subtema_id: subtema._id,
+                                subtema_titulo: subtema.titulo,
+                                informe_id: informeId,
+                                createdAt: hallazgo.createdAt,
+                                updatedAt: hallazgo.updatedAt
+                            });
+                        }
+                    });
+                }
+            });
+        });
+
+        return hallazgos;
     }
 }
