+++ src/tema/tema.service.ts
@@ -84,41 +84,6 @@ export class TemaService {
             .exec();
     }
 
-    async getHallazgosByInforme(informeId: string): Promise<any[]> {
-        const temas = await this.TemaModel
-            .find({ informe_id: informeId, activo: true })
-            .exec();
-
-        const hallazgos = [];
-
-        temas.forEach(tema => {
-            tema.subtema.forEach(subtema => {
-                if (subtema.activo) {
-                    subtema.hallazgo.forEach(hallazgo => {
-                        if (hallazgo.activo) {
-                            hallazgos.push({
-                                _id: hallazgo._id,
-                                titulo: hallazgo.titulo,
-                                criterio: hallazgo.criterio,
-                                descripcion: hallazgo.descripcion,
-                                activo: hallazgo.activo,
-                                tema_id: tema._id,
-                                tema_titulo: tema.titulo,
-                                subtema_id: subtema._id,
-                                subtema_titulo: subtema.titulo,
-                                informe_id: informeId,
-                                createdAt: hallazgo.createdAt,
-                                updatedAt: hallazgo.updatedAt
-                            });
-                        }
-                    });
-                }
-            });
-        });
-
-        return hallazgos;
-    }
-
     async agregarSubtema(temaId: string, SubtemaDTO: SubtemaDTO): Promise<Tema> {
         const tema = await this.TemaModel.findById(temaId).exec();
         if (!tema) {
